Rename and reorganise project files
