+++ src/Generator/Parsing/ParseData.cs
@@ -1,11 +1,7 @@
-﻿using GeneratorV2.Data;
-using System;
+﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace GeneratorV2.Parsing
+namespace Generator.Parsing
 {
     public enum GLAPI
     {
@@ -80,6 +76,47 @@ namespace GeneratorV2.Parsing
         Sync,
         DisplayList,
     }
+
+    public record Expression;
+
+    public record Constant(int Value) : Expression;
+
+    public record CompSize(Expression[] Parameters) : Expression;
+
+    public enum BinaryOperator
+    {
+        Invalid,
+        Addition,
+        Subtraction,
+        Multiplication,
+        Division,
+    }
+
+    public record BinaryOperation(
+        Expression Left,
+        BinaryOperator Operator,
+        Expression Right) : Expression
+    {
+        public static BinaryOperator Invert(BinaryOperator @operator) => @operator switch
+        {
+            BinaryOperator.Addition => BinaryOperator.Subtraction,
+            BinaryOperator.Subtraction => BinaryOperator.Addition,
+            BinaryOperator.Multiplication => BinaryOperator.Division,
+            BinaryOperator.Division => BinaryOperator.Multiplication,
+            _ => throw new Exception("Invalid binary operator, we can't invert it."),
+        };
+
+        public static char GetOperationChar(BinaryOperator @operator) => @operator switch
+        {
+            BinaryOperator.Addition => '+',
+            BinaryOperator.Subtraction => '-',
+            BinaryOperator.Multiplication => '*',
+            BinaryOperator.Division => '/',
+            _ => throw new Exception("Invalid binary operator, there is no char associated."),
+        };
+    }
+
+    public record ParameterReference(string ParameterName) : Expression;
 
     public record GLType();
