+++ src/Generator/Utility/ArrayUtil.cs
@@ -1,10 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace GeneratorV2
+namespace Generator.Utility
 {
     static class ArrayUtil
     {
