+++ src/Generator/Utility/Extensions/DictionaryExtensions.cs
@@ -1,10 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Collections.Generic;
 
-namespace GeneratorV2.Extensions
+namespace Generator.Utility.Extensions
 {
     static class DictionaryExtensions
     {
