+++ src/Generator/Utility/IndentedTextWriter.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace GeneratorV2
+namespace Generator.Utility
 {
     public class IndentedTextWriter : IDisposable
     {
