+++ src/Generator/Utility/Logger.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Runtime.CompilerServices;
 using System.Text;
 
-namespace GeneratorV2
+namespace Generator.Utility
 {
     public class Logger : IDisposable
     {
